+++ src/Nelux/python/Bindings.cpp
@@ -30,7 +30,7 @@ Backend backendFromString(const std::string& backend_str)
 PYBIND11_MODULE(_nelux, m)
 {
     m.doc() = "nelux – lightspeed video decoding into tensors";
-    m.attr("__version__") = "0.12.2";
+    m.attr("__version__") = "0.12.4";
 
     // Expose CUDA build status
 #ifdef NELUX_ENABLE_CUDA